Use debian 11 version of percona-xtrabackup
The percona-xtrabackup-80 package depends on
a package that is not avaialable in the debian 11 repositories, hence the
image build trying to install this package on debian 11 (bullseye) always fails.

Percona provides this package via regular downloads for debian 11, so we use that
to install the compatible package for a successful build.